ajusta calculos de horas do gerador

+++ backoffice/Application/DTOs/Cadastros/Gerador/ViewModel/GeradorViewModel.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string NomeGerador { get; set; }
-        public long QuantidadeHoras { get; set; }
+        public decimal QuantidadeHoras { get; set; }
         public DateTime DataUltimaTrocaOleo { get; set; }
         public long QuantidadeHorasTroca { get; set; }
         public IEnumerable<string>? Checklist { get; set; }
